+++ src/c/utests/data/data.c
@@ -4057,11 +4057,33 @@ static void test_data_const_ui64 (void)
   iot_data_free (data);
 }
 
+static void test_data_const_i64 (void)
+{
+  static iot_data_static_t block;
+  iot_data_t * data = iot_data_alloc_const_i64 (&block, -222);
+  CU_ASSERT (iot_data_i64 (data) == -222)
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
+  CU_ASSERT (iot_data_is_static (data))
+  iot_data_free (data);
+  iot_data_free (data);
+}
+
 static void test_data_const_ui32 (void)
 {
   static iot_data_static_t block;
   iot_data_t * data = iot_data_alloc_const_ui32 (&block, 111u);
   CU_ASSERT (iot_data_ui32 (data) == 111u)
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
+  CU_ASSERT (iot_data_is_static (data))
+  iot_data_free (data);
+  iot_data_free (data);
+}
+
+static void test_data_const_i32 (void)
+{
+  static iot_data_static_t block;
+  iot_data_t * data = iot_data_alloc_const_i32 (&block, -111);
+  CU_ASSERT (iot_data_i32 (data) == -111)
   CU_ASSERT (data == IOT_DATA_STATIC (&block))
   CU_ASSERT (iot_data_is_static (data))
   iot_data_free (data);
@@ -4079,11 +4101,33 @@ static void test_data_const_ui16 (void)
   iot_data_free (data);
 }
 
+static void test_data_const_i16 (void)
+{
+  static iot_data_static_t block;
+  iot_data_t * data = iot_data_alloc_const_i16 (&block, -444);
+  CU_ASSERT (iot_data_i16 (data) == -444)
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
+  CU_ASSERT (iot_data_is_static (data))
+  iot_data_free (data);
+  iot_data_free (data);
+}
+
 static void test_data_const_ui8 (void)
 {
   static iot_data_static_t block;
   iot_data_t * data = iot_data_alloc_const_ui8 (&block, 88u);
   CU_ASSERT (iot_data_ui8 (data) == 88u)
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
+  CU_ASSERT (iot_data_is_static (data))
+  iot_data_free (data);
+  iot_data_free (data);
+}
+
+static void test_data_const_i8 (void)
+{
+  static iot_data_static_t block;
+  iot_data_t * data = iot_data_alloc_const_i8 (&block, -88);
+  CU_ASSERT (iot_data_i8 (data) == -88)
   CU_ASSERT (data == IOT_DATA_STATIC (&block))
   CU_ASSERT (iot_data_is_static (data))
   iot_data_free (data);
@@ -4136,7 +4180,6 @@ static void test_data_const_list (void)
   iot_data_free (list_non_static);
   iot_data_free (list_non_static);
 }
-
 
 static void test_data_const_types (void)
 {
@@ -4864,8 +4907,28 @@ static void test_shallow_copy_map (void)
   CU_ASSERT (copy != NULL)
   CU_ASSERT (copy != map)
   CU_ASSERT (iot_data_type (copy) == IOT_DATA_MAP)
+  CU_ASSERT (iot_data_map_type (copy) == IOT_DATA_MULTI)
+  CU_ASSERT (iot_data_map_key_type (copy) == IOT_DATA_STRING)
   CU_ASSERT (iot_data_equal (copy, map))
   iot_data_map_iter_t iter;
+  iot_data_map_iter (map, &iter);
+  while (iot_data_map_iter_next (&iter))
+  {
+    CU_ASSERT (iot_data_map_iter_value (&iter) == iot_data_map_get (copy, iot_data_map_iter_key (&iter)))
+  }
+  iot_data_free (copy);
+  iot_data_free (map);
+  map = iot_data_alloc_typed_map (IOT_DATA_UINT16, IOT_DATA_POINTER);
+  iot_data_map_add (map, iot_data_alloc_ui16 (1), iot_data_alloc_pointer (&map, NULL));
+  iot_data_map_add (map, iot_data_alloc_ui16 (2), iot_data_alloc_pointer (&copy, NULL));
+  iot_data_map_add (map, iot_data_alloc_ui16 (3), iot_data_alloc_pointer (&iter, NULL));
+  copy = iot_data_shallow_copy (map);
+  CU_ASSERT (copy != NULL)
+  CU_ASSERT (copy != map)
+  CU_ASSERT (iot_data_type (copy) == IOT_DATA_MAP)
+  CU_ASSERT (iot_data_map_type (copy) == IOT_DATA_POINTER)
+  CU_ASSERT (iot_data_map_key_type (copy) == IOT_DATA_UINT16)
+  CU_ASSERT (iot_data_equal (copy, map))
   iot_data_map_iter (map, &iter);
   while (iot_data_map_iter_next (&iter))
   {
@@ -5128,20 +5191,20 @@ static void test_binary_take (void)
 
   // Buffer taken from binary
   bin = iot_data_alloc_binary (buff, sizeof (buff), IOT_DATA_COPY);
-  alloced = (void *) iot_data_address (bin);
+  const void * addr =iot_data_address (bin);
   ptr = iot_data_binary_take (bin, &len);
   CU_ASSERT (len == 4)
-  CU_ASSERT (ptr == alloced)
+  CU_ASSERT (ptr == addr)
   iot_data_free (bin);
   free (ptr);
 
   // Buffer not taken as binary ref counted
   bin = iot_data_alloc_binary (buff, sizeof (buff), IOT_DATA_COPY);
   iot_data_add_ref (bin);
-  alloced = (void *) iot_data_address (bin);
+  addr = iot_data_address (bin);
   ptr = iot_data_binary_take (bin, &len);
   CU_ASSERT (len == 4)
-  CU_ASSERT (ptr != alloced)
+  CU_ASSERT (ptr != addr)
   iot_data_free (bin);
   iot_data_free (bin);
   free (ptr);
@@ -5190,13 +5253,98 @@ static void test_data_tags (void)
 static void test_data_block (void)
 {
   uint32_t size = iot_data_block_size ();
-  uint8_t * ptr = iot_data_block_alloc ();
+  uint8_t * ptr = iot_data_block_alloc (size);
   for (uint8_t i = 0; i < size; i++) // Valgrind check
   {
     ptr[i] = 0u;
   }
   CU_ASSERT (size > 0)
   iot_data_block_free (ptr);
+  ptr = iot_data_block_alloc (size + 1);
+  CU_ASSERT_PTR_NULL (ptr)
+}
+
+static void test_data_iter (void)
+{
+  iot_data_iter_t iter;
+
+  uint32_t vec_size = 5u;
+  uint32_t vec_i = 0u;
+  iot_data_t *vector_data = iot_data_alloc_vector (vec_size);
+  for (uint32_t  i = 0u; i < vec_size; i++)
+  {
+    iot_data_vector_add (vector_data, i, iot_data_alloc_ui32 (i));
+  }
+
+  iot_data_iter (vector_data, &iter);
+  CU_ASSERT_TRUE (iot_data_iter_has_next (&iter))
+  while (iot_data_iter_next (&iter))
+  {
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), vec_i)
+    vec_i++;
+  }
+  while (iot_data_iter_prev (&iter))
+  {
+    vec_i--;
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), vec_i)
+  }
+  iot_data_free (vector_data);
+
+  uint32_t list_size = 5u;
+  uint32_t list_i = 0u;
+  iot_data_t * list_data = iot_data_alloc_list ();
+  for (uint32_t  i = 0u; i < list_size; i++)
+  {
+    iot_data_list_head_push (list_data, iot_data_alloc_ui32 (i));
+  }
+
+  iot_data_iter (list_data, &iter);
+  CU_ASSERT_TRUE (iot_data_iter_has_next (&iter))
+  while (iot_data_iter_next (&iter))
+  {
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), list_i)
+    list_i++;
+  }
+  while (iot_data_iter_prev (&iter))
+  {
+    list_i--;
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), list_i)
+  }
+  iot_data_free (list_data);
+
+  uint32_t  map_size = 5u;
+  uint32_t map_i = 0u;
+  iot_data_t *map_data = iot_data_alloc_map (IOT_DATA_UINT32);
+  for (uint32_t i = 0u; i < map_size; i++)
+  {
+    iot_data_map_add (map_data, iot_data_alloc_ui32 (i), iot_data_alloc_ui32 (i));
+  }
+
+  iot_data_iter (map_data, &iter);
+  CU_ASSERT_TRUE (iot_data_iter_has_next (&iter))
+  while (iot_data_iter_next (&iter))
+  {
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), map_i)
+    map_i++;
+  }
+  while (iot_data_iter_prev (&iter))
+  {
+    map_i--;
+    const iot_data_t *ele = iot_data_iter_value (&iter);
+    CU_ASSERT_PTR_NOT_NULL_FATAL (ele)
+    CU_ASSERT_EQUAL (iot_data_ui32 (ele), map_i)
+  }
+  iot_data_free (map_data);
 }
 
 void cunit_data_test_init (void)
@@ -5348,6 +5496,10 @@ void cunit_data_test_init (void)
   CU_add_test (suite, "data_const_ui32", test_data_const_ui32);
   CU_add_test (suite, "data_const_ui18", test_data_const_ui16);
   CU_add_test (suite, "data_const_ui8", test_data_const_ui8);
+  CU_add_test (suite, "data_const_i64", test_data_const_i64);
+  CU_add_test (suite, "data_const_i32", test_data_const_i32);
+  CU_add_test (suite, "data_const_i18", test_data_const_i16);
+  CU_add_test (suite, "data_const_i8", test_data_const_i8);
   CU_add_test (suite, "data_const_pointer", test_data_const_pointer);
   CU_add_test (suite, "data_const_list", test_data_const_list);
   CU_add_test (suite, "data_const_types", test_data_const_types);
@@ -5384,4 +5536,5 @@ void cunit_data_test_init (void)
   CU_add_test (suite, "data_is_nan", test_data_is_nan);
   CU_add_test (suite, "data_tags", test_data_tags);
   CU_add_test (suite, "data_block", test_data_block);
+  CU_add_test (suite, "data_iter", test_data_iter);
 }